lib/tgapi: Fix not responding bot when in a private chat
Dwiky reports that the bot is not responding the ping command in a
private chat. This happens because the tgj_get_chat() function returns
-ENOENT when the chat object doesn't have the "title" field. This
behavior is wrong because in a private chat, there is no
"message.chat.title" field.

Fix it by making the "title" field optional. This field is only
available on a group chat.

Link: https://t.me/GNUWeeb/698314
Fixes: 593f7d012d5082b6c060c7e96b3d1b1ccb21ff1d ("Initial commit")
Reported-by: Dwiky Rizky Ananditya <kyzsuki@gnuweeb.org>
Tested-by: Dwiky Rizky Ananditya <kyzsuki@gnuweeb.org>
Acked-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>